Applied review suggestion, refactoring
