+++ hermes-common/src/main/java/pl/allegro/tech/hermes/common/kafka/KafkaHeaderNameParameters.java
@@ -1,0 +1,13 @@
+package pl.allegro.tech.hermes.common.kafka;
+
+public interface KafkaHeaderNameParameters {
+
+    String getSchemaVersion();
+
+    String getSchemaId();
+
+    String getMessageId();
+
+    String getTimestamp();
+
+}
