--- hermes-frontend/src/main/java/pl/allegro/tech/hermes/frontend/producer/kafka/KafkaHeaderNameParameters.java
@@ -1,12 +1,0 @@
-package pl.allegro.tech.hermes.frontend.producer.kafka;
-
-public interface KafkaHeaderNameParameters {
-
-    String getMessageId();
-
-    String getTimestamp();
-
-    String getSchemaVersion();
-
-    String getSchemaId();
-}
